requirements.txt
fastapi >= 0.111.0
requests >= 2.32.3
beautifulsoup4 >= 4.12.3
python-dotenv >= 1.0.1
pydantic>=2.7.0
mysql-connector-python